Refactor relay -> parachain XCM sending function
