+++ packages/shared/src/helpers/index.ts
@@ -1,6 +1,8 @@
 import { withExpect } from '@acala-network/chopsticks-testing'
 import { assert, expect } from 'vitest'
 
+import type { StorageValues } from '@acala-network/chopsticks'
+import type { ApiPromise } from '@polkadot/api'
 import type { HexString } from '@polkadot/util/types'
 
 const { check, checkEvents, checkHrmp, checkSystemEvents, checkUmp } = withExpect((x: any) => ({
@@ -62,7 +64,7 @@ export function objectCmp(
  * Given a PJS client and a hex-encoded extrinsic with a given origin, modify the
  * `scheduler` pallet's `agenda` storage to execute the extrinsic in the next block.
  */
-export async function scheduleCallWithOrigin(client: any, call: HexString, origin: any) {
+export async function scheduleCallWithOrigin(client: any, encodedCall: HexString, origin: any) {
   const number = (await client.api.rpc.chain.getHeader()).number.toNumber()
 
   await client.dev.setStorage({
@@ -73,7 +75,7 @@ export async function scheduleCallWithOrigin(client: any, call: HexString, origi
           [
             {
               call: {
-                Inline: call,
+                Inline: encodedCall,
               },
               origin: origin,
             },
@@ -82,4 +84,73 @@ export async function scheduleCallWithOrigin(client: any, call: HexString, origi
       ],
     },
   })
+}
+
+export async function getBlockNumber(client: any): Promise<number> {
+  return (await client.api.rpc.chain.getHeader()).number.toNumber()
+}
+
+/**
+ * Send an XCM message containing an extrinsic to be executed in a parachain with a given origin.
+ *
+ * @param relayClient Relay chain client form which to execute `xcmPallet.send`
+ * @param parachainId ID of the parachain to which the XCM message is to be sent
+ * @param call Hex-encoded identity pallet extrinsic
+ * @param origin Origin with which the extrinsic is to be executed at the location parachain
+ * @param requireWeightAtMost Reftime/proof size parameters that `send::Transact` may require (only in XCM v4);
+ *        sensible defaults are given.
+ */
+export async function xcmSendTransact(
+  relayClient: {
+    api: ApiPromise
+    dev: {
+      setStorage: (values: StorageValues, blockHash?: string) => Promise<any>
+    }
+  },
+  parachainId: number,
+  call: HexString,
+  origin: any,
+  requireWeightAtMost = { proofSize: '10000', refTime: '100000000' },
+): Promise<any> {
+  // Destination of the XCM message sent from the relay chain to the parachain via `xcmPallet`
+  const dest = {
+    V4: {
+      parents: 0,
+      interior: {
+        X1: [
+          {
+            Parachain: parachainId,
+          },
+        ],
+      },
+    },
+  }
+
+  // The message being sent to the parachain, containing a call to be executed in the parachain:
+  const message = {
+    V4: [
+      {
+        UnpaidExecution: {
+          weightLimit: 'Unlimited',
+          checkOrigin: null,
+        },
+      },
+      {
+        Transact: {
+          call: {
+            encoded: call,
+          },
+          originKind: 'SuperUser',
+          requireWeightAtMost,
+        },
+      },
+    ],
+  }
+
+  const xcmTx = relayClient.api.tx.xcmPallet.send(dest, message)
+  const encodedRelayCallData = xcmTx.method.toHex()
+
+  /// See `scheduleCallWithOrigin` for more.
+
+  await scheduleCallWithOrigin(relayClient, encodedRelayCallData, origin)
 }
